:recycle: Refactor Card in its own file
